requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-jose==3.3.0
python-jose[cryptography]
python-dotenv
